HW/7 - Перенес доменные сущности, которые относятся к фильмам в пакет entity
