--- app/src/main/java/ru/subbotind/android/academy/myfirstapp/data/Actor.kt
@@ -1,7 +1,0 @@
-package ru.subbotind.android.academy.myfirstapp.data
-
-data class Actor(
-    val id: Int,
-    val name: String,
-    val picture: String
-)
