--- app/src/main/java/ru/subbotind/android/academy/myfirstapp/data/Genre.kt
@@ -1,3 +1,0 @@
-package ru.subbotind.android.academy.myfirstapp.data
-
-data class Genre(val id: Int, val name: String)
