--- app/src/main/java/ru/subbotind/android/academy/myfirstapp/data/Movie.kt
@@ -1,15 +1,0 @@
-package ru.subbotind.android.academy.myfirstapp.data
-
-data class Movie(
-    val id: Int,
-    val title: String,
-    val overview: String,
-    val poster: String,
-    val backdrop: String,
-    val ratings: Float,
-    val numberOfRatings: Int,
-    val minimumAge: Int,
-    val runtime: Int,
-    val genres: List<Genre>,
-    val actors: List<Actor>
-)
