+++ app/src/main/java/ru/subbotind/android/academy/myfirstapp/domain/entity/Genre.kt
@@ -1,0 +1,6 @@
+package ru.subbotind.android.academy.myfirstapp.domain.entity
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Genre(val id: Int, val name: String)
